ocr_backend/requirements.txt
fastapi>=0.110.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
uvicorn>=0.28.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
Pillow>=10.2.0
rapidfuzz>=3.6.0
groq>=0.4.2
pytest>=8.0.0
httpx>=0.27.0
streamlit>=1.32.0
python-multipart>=0.0.9
